finish

+++ pkg/scheduler/car.go
@@ -5,11 +5,11 @@ type Car struct {
 	carId            int64
 	queueId          int64
 	chargingType     int
-	chargingQuantity int
+	chargingQuantity float64
 	isCharging       bool
 }
 
-func NewCar(userId int64, carId int64, queueId int64, chargingType int, chargingQuantity int) *Car {
+func NewCar(userId int64, carId int64, queueId int64, chargingType int, chargingQuantity float64) *Car {
 	return &Car{userId, carId, queueId, chargingType, chargingQuantity, false}
 
 }
@@ -22,7 +22,7 @@ func (car *Car) GetCarId() int64 {
 	return car.carId
 }
 
-func (car *Car) GetChargingQuantity() int {
+func (car *Car) GetChargingQuantity() float64 {
 	return car.chargingQuantity
 }
